+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/configuration/Configuration.java
@@ -66,9 +66,7 @@ public final class Configuration {
     public static final Level logLevel;
 
     public static boolean showStat;
-    
-    public static final String ABC_DEPS_SEPARATOR = "{*sep*}";
-
+        
     @ConfigurationDefaultBoolean(true)
     @ConfigurationCategory("ui")
     public static ConfigurationItem<Boolean> openMultipleFiles = null;