+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/configuration/CustomConfigurationKeys.java
@@ -20,12 +20,13 @@ package com.jpexs.decompiler.flash.configuration;
  *
  * @author JPEXS
  */
-public class CustomConfigurationKeys {
-
+public class CustomConfigurationKeys {   
+    
     public static final String KEY_LAST_SELECTED_PATH_RESOURCES = "lastSelectedPath.resources";
     public static final String KEY_LAST_SELECTED_PATH_TAGLIST = "lastSelectedPath.taglist";
     public static final String KEY_CHARSET = "charset";
     public static final String KEY_LIBRARY = "library";
     public static final String KEY_LOADED_IMPORT_ASSETS = "loadedImportAssets";
     public static final String KEY_ABC_DEPENDENCIES = "abcDependencies";
+    public static final String KEY_BREAKPOINTS = "breakpoints";
 }